+++ plugins/12_Midnight/localization/enUS.lua
@@ -95,6 +95,9 @@ L['options_icons_spiritpaw_marathon'] = '{achievement:62202}'
 L['options_icons_spiritpaw_marathon_desc'] = 'Display the locations for the {achievement:62202} achievement.'
 L['spiritpaw_marathon_note'] = 'Speak with {npc:258938}, then pick up the nearby {npc:250100} and take them to the {dot:Pink} within the {location:Temple of Jan\'alai} within 30 minutes. \n\n ' .. ns.color.Red('Do not mount up or you will lose the buff.')
 
+L['options_icons_gnome_alone'] = '{achievement:62200}'
+L['options_icons_gnome_alone_desc'] = 'Display the locations for the {achievement:62200} achievement.'
+
 -------------------------------------------------------------------------------
 ---------------------------------- ARCANTINA ----------------------------------
 -------------------------------------------------------------------------------